fixed HotelServicesController

+++ HotelManagerSystem.API/AuthBL/Controllers/DirectoriesController/HotelServicesController.cs
@@ -1,13 +1,18 @@
-﻿using HotelManagerSystem.BL.Directories;
+﻿using HotelManagerSystem.API.Responses;
+using HotelManagerSystem.BL.Directories;
+using HotelManagerSystem.BL.Exceptions;
 using HotelManagerSystem.DAL.Data;
 using HotelManagerSystem.DAL.Responses;
 using HotelManagerSystem.Models.Data;
+using HotelManagerSystem.Models.DTOs;
 using HotelManagerSystem.Models.Entities;
 using HotelManagerSystem.Models.Request.CreateRequest;
 using HotelManagerSystem.Models.Request.UpdateRequest;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Diagnostics.Metrics;
+using System.Linq;
 
 namespace HotelManagerSystem.API.AuthBL.Controllers.DirectoriesController
 {
@@ -28,63 +33,73 @@ namespace HotelManagerSystem.API.AuthBL.Controllers.DirectoriesController
 
         [HttpPost]
         [Route("create")]
-        public async Task<Response> Create([FromBody] CreateDescDirectoryRequest request)
+        public async Task<ActionResult> Create([FromBody] CreateDescDirectoryRequest request)
         {
             try
             {
-                await _service.Create(request);
+                var service = await _service.Create(request);
+                return Created("", service);
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error while processing request from {Hotel Services}", request);
+                return BadRequest(new ErrorResponse(ex.Message));
             }
-
-            return new Response(200, true, null);
         }
 
         [HttpPut]
         [Route("update")]
-        public async Task<Response> Update(UpdateDescDirectoryRequest request)
+        public async Task<ActionResult> Update(UpdateDescDirectoryRequest request)
         {
             try
             {
                 await _service.Update(request);
             }
+            catch (EntityNotFoundException<HotelServices> ex)
+            {
+                return NotFound(new ErrorResponse(ex.Message));
+            }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error while processing request from {Hotel Services}", request);
+                return BadRequest(new ErrorResponse(ex.Message));
             }
 
-            return new Response(200, true, null);
+            return Ok();
         }
 
         [HttpDelete]
         [Route("deleteById")]
-        public async Task<Response> Delete(int id)
+        public async Task<ActionResult> Delete(int id)
         {
+            HotelServices hotelServices = await _service.GetByIdAsync(id);
+            if (hotelServices == null)
+                return NotFound(new ErrorResponse("Hotel Services not found"));
             await _service.Delete(id);
 
-            return new Response(200, true, null);
+            return NoContent();
         }
 
         [HttpGet]
         [Route("GetById")]
         [Authorize]
-        public async Task<HotelServicesResponse> GetById(int id)
+        public async Task<ActionResult<HotelServicesDto>> GetById(int id)
         {
-            HotelServices services = await _service.GetByIdAsync(id);
+            HotelServices hotelServices = await _service.GetByIdAsync(id);
+            if (hotelServices == null)
+                return NotFound(new ErrorResponse("Hotel Services not found"));
 
-            return new HotelServicesResponse(200, true, null, services);
+            return Ok(new HotelServicesDto(hotelServices));
         }
 
         [HttpGet]
         [Route("GetAll")]
         [Authorize]
-        public async Task<HotelServicesListResponse> GetAll()
+        public async Task<ActionResult<List<HotelServicesDto>>> GetAll()
         {
             var list = await _service.GetAll();
 
-            return new HotelServicesListResponse(200, true, null, list);
+            return Ok(list.Select(dep => new HotelServicesDto(dep)).ToList());
         }
     }
 }
